typeDefProducts.js file deledted and intrigrated into typeDef.js
