+++ server/schema/typeDefs.js
@@ -19,6 +19,21 @@ const typeDefs = gql`
   type Mutation {
     addUser(username: String!, email: String!, password: String!): Auth
   }
+##########################
+
+  type Product {
+    _id: ID
+    name: String
+    photo: String
+    description: String
+    price: Number
+  }
+
+  type Query {
+    Products: [Product]
+  }
+#############################
+
 `;
 
 module.exports = typeDefs;
